+++ src/constants.ts
@@ -1,0 +1,2 @@
+
+export const taskSource = 'Vitis HLS IDE';